+++ CHANGES.md
@@ -3,6 +3,7 @@
 #### [unreleased]
 * add customizer background color to login screen.
 * add some subscription helper functions.
+* add shortcodes for family profile display.
 
 #### 1.6.3 / 2017-10-6
 * add custom-logo for login/lost-password page.